Added support for structures.
Structures are complex repeating block arrangements currently placed in 'crops' and 'developed' areas, as a replacement for the old building code.  They aren't perfect but add more realism and Minecraft-relevant content while not impacting speed.

This required the terrain algorithm to place 'End Stone' instead of 'Stone' underground so 'Stone' could be used in structures.  The 'End Stone' is replaced with ores during ore placement, and is finally replaced with 'Stone' when the region is assembled.

createstructures.py: new file!
newore.py: changed stoneID to 'End Stone'
newterrain.py: imported YAML for structures
	       added new structures methods
	       created "simple" non-structures versions of placecrops and placedeveloped
	       added placeice as per previous commit (sigh)
	       added support for 'End Stone'