Minor changes to BuildRegion.py to inform the user
The time spent planting trees and depositing ores on the region
level can be large, and the user often does not know how much
work is being done.  Now the user is informed of how many trees
are to be planted and ores are to be deposited.

BuildRegion.py: new output added