Trapped pyopencl.RuntimeError to check for timeout
The OpenCL feature leverages the GPU for region preparation,
but this can fail if the GPU is being used for rendering a
display.  This failure is usually exhibited by a timeout.
The code now checks for this event and informs the user.

clidt.py: added try/except with print statements