Fixes #17, updates pymclevel, very minor cleanup.
The problem with Alaska and Hawaii maps has been resolved by
converting every map to the target SRS before further processing.

Additionally, I have updated pymclevel to the latest version in order
to use the supplied Height value and some other performance improvements.

I also did some very minor cleanup of the code in buildvrts(), and
handled the nodata-that-isn't-reported-as-nodata case for landcover.