Added initial OpenCL support for elevation.
Got to start somewhere.  Replaced the numpy version of shoehorning of
the elevation array into Minecraft dimensions with OpenCL.  It's
currently twenty times slower, but it's just a base for the eventual
replacement of the gdalwarp code.

Modified idt.py to resize the output as appropriate.  Changed
region.py and crust.py to match.  Also fussed with crust a little to
make it more PEP 8 compliant and to remove unnecessary objects.

Both idt.py and elev.py have support for generating and using test
objects.  They raise AssertionError when errors exceed 1%.

Moved chunks to utils.py since multiple modules now use it.