First pass at supporting new tiled files, fixing most of #25.
The software works again.  It now downloads tiled files from the USGS.
These files are now cached since they are very large and far more than
the typical user will need for their purposes.  While the code works,
it is extremely inefficient, especially in PrepRegion.py.

A minor change to GetRegion.py corrects a print statement while two
lines changed in terrain.py to reflect the new landcover product IDs.
The rest of the changes were in region.py.  New product IDs for
elevation and landcover were required.  The product availability code
now checks to see if the products are tiled instead of seamless or
read only (neither of which are usable now) and all the old layerID
code is obsolete.  The download file routine has changed significantly
and is less hackish.

Things that need to be added before the hotfix is complete:
 * Map files need to be pre-shrunk before PrepRegion.py is called.
 * Progress bars are needed for downloads and possibly elsewhere.
 * Old code needs to be stripped out.