Download resume and progress bars added, some old code removed.  References #25.
Removed decodeLayerID() because that layer ID scheme is obsolete, and
removed buildvrts() because the current approach works.

The USGS web server does not handle range checks across redirected
files.  So now we check whether we've reached the end of the redirect
chain before applying the range check.  Once we have the range check,
the USGS web server promptly ignores the request and sends us the
entire file anyway, so we drop the unneeded data on the floor.  This
is still better than deleting the file and starting from scratch
because if the USGS web server ever fixes the Range request behavior,
we will be able to take advantage of it immediately.  The code that
checks for the bugfix has not yet been written.

Accidentally left in some bad code here: there's a redirect to use
gdalwarp on the landcover files but for some reason that generates bad
value 67 on BlockIsland.  I really need to fix the ReadAsArray call!