Landcover files are now optimized, with other minor changes.  References #25.
Only the required portion of the landcover file is accessed now, which
saves time and memory.  The VRTs are warped in advance -- this is
faster than warping individual input files.

The files we create and store are now referenced via absolute paths.

In addition, I abstracted out the 'layername' functionality into a
method.  There is probably a better name for this but I can't think of
it.  Given a layer ID, tells what type of layer it is.