Initial attempts to work with updated pymclevel.
Ubuntu's packages of GDAL are a little wacky, but I found and
documented a workaround.  Added virtualenv to the instructions.  Roses
are out, poppies are in.  Crops are out, wheat is in.  Leaves and wood
are no longer generic.

There are other significant changes internal to pymclevel, and it may
be necessary or prudent to rewrite TopoMC due to the magnitude of
these changes.

References #44